MCS-644 removed util package and moved functions to other classes
