+++ machine_common_sense/__init__.py
@@ -11,6 +11,7 @@ from .action import Action
 from .controller import Controller
 from .goal_metadata import GoalMetadata, GoalCategory
 from .material import Material
+from .numeric import Numeric
 from .object_metadata import ObjectMetadata
 from .pose import Pose
 from .return_status import ReturnStatus
@@ -18,8 +19,8 @@ from .reward import Reward
 from .scene_history import SceneHistory
 from .history_writer import HistoryWriter
 from .step_metadata import StepMetadata
-from .util import Util
 from .serializer import SerializerMsgPack, SerializerJson
+from .to_string import ToString
 from ._version import __version__
