+++ machine_common_sense/controller.py
@@ -29,6 +29,8 @@ DEFAULT_MOVE = 0.1
 
 from .action import Action
 from .goal_metadata import GoalMetadata
+from .material import Material
+from .numeric import Numeric
 from .object_metadata import ObjectMetadata
 from .plotter import TopDownPlotter
 from .pose import Pose
@@ -38,7 +40,7 @@ from .scene_history import SceneHistory
 from .step_metadata import StepMetadata
 from .recorder import VideoRecorder
 from .uploader import S3Uploader
-from .util import Util
+from .to_string import ToString
 from .history_writer import HistoryWriter
 from .config_manager import ConfigManager
 
@@ -588,7 +590,7 @@ class Controller():
         receptacleObjectImageCoordsY = kwargs.get(
             self.RECEPTACLE_IMAGE_COORDS_Y_KEY, self.DEFAULT_IMG_COORD)
 
-        if not Util.is_number(amount, self.AMOUNT_KEY):
+        if not Numeric.is_number(amount, self.AMOUNT_KEY):
             # The default for open/close is 1, the default for "Move" actions
             # is 0.5
             if action in self.OBJECT_MOVE_ACTIONS:
@@ -596,38 +598,38 @@ class Controller():
             else:
                 amount = self.DEFAULT_AMOUNT
 
-        if not Util.is_number(force, self.FORCE_KEY):
+        if not Numeric.is_number(force, self.FORCE_KEY):
             force = self.DEFAULT_FORCE
 
         # Check object directions are numbers
-        if not Util.is_number(
+        if not Numeric.is_number(
                 objectImageCoordsX,
                 self.OBJECT_IMAGE_COORDS_X_KEY):
             objectImageCoordsX = self.DEFAULT_IMG_COORD
 
-        if not Util.is_number(
+        if not Numeric.is_number(
                 objectImageCoordsY,
                 self.OBJECT_IMAGE_COORDS_Y_KEY):
             objectImageCoordsY = self.DEFAULT_IMG_COORD
 
         # Check receptacle directions are numbers
-        if not Util.is_number(
+        if not Numeric.is_number(
                 receptacleObjectImageCoordsX,
                 self.RECEPTACLE_IMAGE_COORDS_X_KEY):
             receptacleObjectImageCoordsX = self.DEFAULT_IMG_COORD
 
-        if not Util.is_number(
+        if not Numeric.is_number(
                 receptacleObjectImageCoordsY,
                 self.RECEPTACLE_IMAGE_COORDS_Y_KEY):
             receptacleObjectImageCoordsY = self.DEFAULT_IMG_COORD
 
-        amount = Util.is_in_range(
+        amount = Numeric.is_in_range(
             amount,
             self.MIN_AMOUNT,
             self.MAX_AMOUNT,
             self.DEFAULT_AMOUNT,
             self.AMOUNT_KEY)
-        force = Util.is_in_range(
+        force = Numeric.is_in_range(
             force,
             self.MIN_FORCE,
             self.MAX_FORCE,
@@ -673,14 +675,15 @@ class Controller():
         teleportRotation = None
         teleportPosition = None
 
-        if teleportRotInput is not None and Util.is_number(teleportRotInput):
+        if teleportRotInput is not None and Numeric.is_number(
+                teleportRotInput):
             teleportRotation = {}
             teleportRotation['y'] = kwargs.get(self.TELEPORT_Y_ROT)
 
         if (teleportPosXInput is not None and
-                Util.is_number(teleportPosXInput) and
+                Numeric.is_number(teleportPosXInput) and
                 teleportPosZInput is not None and
-                Util.is_number(teleportPosZInput)):
+                Numeric.is_number(teleportPosZInput)):
             teleportPosition = {}
             teleportPosition['x'] = teleportPosXInput
             teleportPosition['z'] = teleportPosZInput
@@ -730,7 +733,7 @@ class Controller():
             return None
 
         if ',' in action:
-            action, kwargs = Util.input_to_action_and_params(action)
+            action, kwargs = Action.input_to_action_and_params(action)
 
         action_list = self.retrieve_action_list_at_step(
             self._goal,
@@ -963,7 +966,7 @@ class Controller():
         action_list = goal_config.get('action_list', [])
         for index, action_list_at_step in enumerate(action_list):
             action_list[index] = [
-                Util.input_to_action_and_params(action)
+                Action.input_to_action_and_params(action)
                 if isinstance(action, str) else action
                 for action in action_list_at_step
             ]
@@ -1029,7 +1032,7 @@ class Controller():
         material_list = (
             list(
                 filter(
-                    Util.verify_material_enum_string,
+                    Material.verify_material_enum_string,
                     [
                         material.upper()
                         for material in object_metadata['salientMaterials']
@@ -1325,7 +1328,7 @@ class Controller():
             "OBJECTS: " + str(len(restricted_output.object_list)) + " TOTAL"
         )
         if len(restricted_output.object_list) > 0:
-            for line in Util.generate_pretty_object_output(
+            for line in ToString.generate_pretty_object_output(
                 restricted_output.object_list
             ):
                 logger.debug("    " + line)
