+++ scripts/kyle-test.ini
@@ -1,0 +1,6 @@
+[MCS]
+history_enabled: true
+metadata: level1
+save_debug_images: false
+save_debug_json: true
+video_enabled: false
